+++ Elements/src/Model.cs
@@ -250,7 +250,7 @@ namespace Elements
                 return elements;
             }
 
-            var props = t.GetProperties(BindingFlags.Public | BindingFlags.Instance);
+            var props = t.GetProperties(BindingFlags.Public | BindingFlags.Instance).Where(p => !p.GetCustomAttributes<JsonIgnoreAttribute>().Any());
             var constrainedProps = props.Where(p => IsValidForRecursiveAddition(p.PropertyType));
             foreach (var p in constrainedProps)
             {
@@ -372,6 +372,18 @@ namespace Elements
                 elements.AddRange(model.AllElementsOfType<T>());
             }
             return elements;
+        }
+
+        /// <summary>
+        /// Get all proxies of a certain type from a specific model name in a dictionary of models.
+        /// </summary>
+        /// <param name="models">Dictionary of models keyed by string</param>
+        /// <param name="modelName">The name of the model</param>
+        /// <typeparam name="T">The type of element we want to retrieve</typeparam>
+        /// <returns></returns>
+        public static List<ElementProxy<T>> AllProxiesOfType<T>(this Dictionary<string, Model> models, string modelName) where T : Element
+        {
+            return models.AllElementsOfType<T>(modelName).Proxies(modelName).ToList();
         }
     }
 }