+++ CHANGELOG.md
@@ -16,10 +16,13 @@
 ### Changed
 
 - `Identity.AddOverrideIdentity` is now an extension method.
+- Profile operations throw fewer exceptions when some piece of the profile is invalid, preferring instead to return a partial result or a null.
 
 ### Fixed
 
 `Line.ExtendTo` would sometimes return erroneous results if any of the trimming segments crossed the origin.
+
+### Fixed
 
 ## 0.9.3